extract request types
