+++ src/components/Playlists/hooks/useTrackList.ts
@@ -7,12 +7,7 @@ import {
 } from 'react';
 import { useRecoilState, useSetRecoilState } from 'recoil';
 
-import {
-  PagedResponse,
-  SpotifyPlaylistTrack,
-  StoredSpotifyPlaylistTrack,
-  TrackState,
-} from '../../../types';
+import { SpotifyPlaylistTrack, StoredSpotifyPlaylistTrack } from '../../../types';
 import { SPOTIFY_PLAYLIST_TRACKS } from '../../../consts';
 import { getPlaylistTracks } from '../../../utils/externalData';
 import extractTrackData from '../../../utils/extractTrackData';
