+++ src/types/spotify-revised/common.d.ts
@@ -4,6 +4,11 @@ declare global {
     ALBUM = 'ALBUM',
     ARTIST = 'ARTIST',
   }
+
+  export enum TrackState {
+    VALID = 'VALID',
+    INVALID = 'INVALID',
+  }
 }
 
 export {};
