+++ src/utils/authRequest.ts
@@ -1,8 +1,6 @@
 import { getValidToken } from '../components/Auth/utils';
 import { get as getPlain, post as postPlain, getPaged as getPagedPlain } from './request';
 
-import { ContentType, PostData, QueryStringData } from '../types';
-
 function parseResponse<S>(response: Response): Promise<S> {
   const contentType = response.headers.get('content-type');
