+++ src/utils/request.ts
@@ -1,10 +1,3 @@
-import {
-  ContentType,
-  PagedResponse,
-  PostData,
-  QueryStringData,
-} from '../types';
-
 export function urlWithQueryString(url: string, data?: QueryStringData): string {
   if (!data || Object.keys(data).length === 0) {
     return url;
